Move wait for slot function to top level
